add useWindowSize hook
